Added default methods for stripe models, more need to be added Changed <case class>.tupled to (<case class>.apply _).tupled so it works with companion objects Updated read to describe the default methods Changed expMonth/expYear from Long to Int

+++ src/main/scala/org/mdedetrich/stripe/v1/Sources.scala
@@ -7,9 +7,9 @@ package org.mdedetrich.stripe.v1
 object Sources {
 
   trait BaseCardSource {
-    def expMonth: Long
+    def expMonth: Int
 
-    def expYear: Long
+    def expYear: Int
 
     def number: String
